requirements.txt
streamlit>=1.57.0
plotly>=6.7.0
pandas>=3.0.3
numpy>=2.4.4
scipy>=1.17.1
scikit-learn>=1.5.0
umap-learn>=0.5.6
yfinance>=0.2.40
fredapi>=0.5.2
python-dotenv>=1.0.1
requests>=2.33.1
